requirements.txt
google-genai
redis
yfinance
feedparser
python-telegram-bot
fastapi
uvicorn
jinja2
python-dotenv
requests
